+++ include/kmerhash/experimental/hashmap_robinhood_doubling_offsets.hpp
@@ -440,7 +440,7 @@ protected:
 
 #if defined(REPROBE_STAT)
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 
 		return std::make_pair(i, result);
@@ -520,7 +520,7 @@ protected:
 			reprobe = i - pos;
 		}
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 		return i;
 
@@ -612,7 +612,7 @@ protected:
 			reprobe = i - pos;
 		}
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 		return i;
