+++ include/kmerhash/experimental/hashmap_robinhood_doubling_offsets2.hpp
@@ -445,7 +445,7 @@ protected:
 
 #if defined(REPROBE_STAT)
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 
 		return std::make_pair(i, result);
@@ -525,7 +525,7 @@ protected:
 			reprobe = i - pos;
 		}
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 		return i;
 
@@ -617,7 +617,7 @@ protected:
 			reprobe = i - pos;
 		}
 		this->reprobes += reprobe;
-		this->max_reprobes = std::max(this->max_reprobes, reprobe);
+		this->max_reprobes = std::max(this->max_reprobes, static_cast<size_t>(reprobe));
 #endif
 		return i;
